+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Exclamation symbols (:exclamation:) note something of importance e.g. breaking c
 ### Fixed
 - Entity relations and wrong types for payments.
 - Allow empty string for `switch_inline_query` and `switch_inline_query_current_chat` (InlineKeyboardButton).
+- Fix empty date entry for User and Chat entities, using the current timestamp instead.
 ### Security
 
 ## [0.51.0] - 2017-12-05